Implement Community controller and service

+++ src/main/java/iyteyazilim/projects/digitalcard/repository/ICommunityRepository.java
@@ -3,5 +3,5 @@ package iyteyazilim.projects.digitalcard.repository;
 import iyteyazilim.projects.digitalcard.entity.Community;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ICommunityRepository extends JpaRepository<Community, Integer> {
+public interface ICommunityRepository extends JpaRepository<Community, Long> {
 }
